+++ lib/rspec/core/rake_task.rb
@@ -1,4 +1,3 @@
-require 'rspec/core/backward_compatibility'
 require 'rspec/core/deprecation'
 require 'rake'
 require 'rake/tasklib'