+++ spec/rspec/core/formatters/base_text_formatter_spec.rb
@@ -145,50 +145,6 @@ describe RSpec::Core::Formatters::BaseTextFormatter do
         end
       end
     end
-
-    context 'for #share_as' do
-      before { allow(RSpec).to receive(:deprecate) }
-
-      it 'outputs the name and location' do
-
-        group.share_as :FooBar do
-          it("example name") { expect("this").to eq("that") }
-        end
-
-        line = __LINE__.next
-        group.send(:include, FooBar)
-
-        run_all_and_dump_failures
-
-        expect(output.string).to include(
-          'Shared Example Group: "FooBar" called from ' +
-            "./spec/rspec/core/formatters/base_text_formatter_spec.rb:#{line}"
-        )
-      end
-
-      context 'that contains nested example groups' do
-        it 'outputs the name and location' do
-
-          group.share_as :NestedFoo do
-            describe 'nested group' do
-              describe 'hell' do
-                it("example name") { expect("this").to eq("that") }
-              end
-            end
-          end
-
-          line = __LINE__.next
-          group.send(:include, NestedFoo)
-
-          run_all_and_dump_failures
-
-          expect(output.string).to include(
-            'Shared Example Group: "NestedFoo" called from ' +
-              "./spec/rspec/core/formatters/base_text_formatter_spec.rb:#{line}"
-          )
-        end
-      end
-    end
   end
 
   describe "#dump_pending" do
@@ -279,50 +235,6 @@ describe RSpec::Core::Formatters::BaseTextFormatter do
 
             expect(output.string).to include(
               'Shared Example Group: "foo bar" called from ' +
-              "./spec/rspec/core/formatters/base_text_formatter_spec.rb:#{line}"
-            )
-          end
-        end
-      end
-
-      context 'for #share_as' do
-        before { allow(RSpec).to receive(:deprecate) }
-
-        it 'outputs the name and location' do
-
-          group.share_as :FooBar2 do
-            it("example name") { pending { expect("this").to eq("that") } }
-          end
-
-          line = __LINE__.next
-          group.send(:include, FooBar2)
-
-          run_all_and_dump_pending
-
-          expect(output.string).to include(
-            'Shared Example Group: "FooBar2" called from ' +
-            "./spec/rspec/core/formatters/base_text_formatter_spec.rb:#{line}"
-          )
-        end
-
-        context 'that contains nested example groups' do
-          it 'outputs the name and location' do
-
-            group.share_as :NestedFoo2 do
-              describe 'nested group' do
-                describe 'hell' do
-                  it("example name") { pending { expect("this").to eq("that") } }
-                end
-              end
-            end
-
-            line = __LINE__.next
-            group.send(:include, NestedFoo2)
-
-            run_all_and_dump_pending
-
-            expect(output.string).to include(
-              'Shared Example Group: "NestedFoo2" called from ' +
               "./spec/rspec/core/formatters/base_text_formatter_spec.rb:#{line}"
             )
           end